Committing checkpoint of cleanup/bugfix work
Very much still a hacky prototype, but this seems to work a little more
robustly. Most notably, we no longer make any assumptions about the
Mozilla window-- we just warp the pointer and get the information
that way (this is the method Google Chrome seems to use for its UI
automation tests on Linux).